refactor: move parseLocalLookupResources to common.go
